added export of GOPATH

+++ install/1_clean_code.sh
@@ -10,6 +10,6 @@ case $yn in
 esac
 
 # removing data from chain
-echo "removing $ETHPATH"
-chmod -R 777 $ETHPATH
-rm -rf $ETHPATH
+echo "removing $ETHPATH (if exists)"
+if [ -d "$WORKING_DIR" ]; then chmod -R 777 $ETHPATH; fi
+if [ -d "$WORKING_DIR" ]; then rm -rf $ETHPATH; fi
